Online_Courses_App/online_courses_app/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.52.0
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
oauthlib==3.2.2
phonenumbers==8.13.4
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
zipp==3.12.0
